Fix several minor bugs reported by Claude Opus

+++ src/resqml2/AbstractSeismicLineFeature.cpp
@@ -166,9 +166,9 @@ void AbstractSeismicLineFeature::setTraceLabels(const std::vector<std::string> &
 	uint64_t strNb = values.size();
 	std::unique_ptr<unsigned char[]> cTab(new unsigned char[strNb*dimTwo]);
 
-	int indStr = 0;
+	size_t indStr = 0;
 	for (std::vector<std::string>::const_iterator it = values.begin(); it != values.end(); ++it) {
-		for (unsigned int indCha = 0; indCha < dimTwo; indCha++) {
+		for (uint64_t indCha = 0; indCha < dimTwo; indCha++) {
 			if (indCha < (*it).length())
 				cTab[(indStr*dimTwo) + indCha] = (*it)[indCha];
 			else
